Transitioning to Tab list UI element
