+++ stocksense-app/webapp/pages/playground/__init__.py
@@ -2,4 +2,4 @@ from .data import data
 from .index import playground
 from .strategy import strategy
 
-__all__ = ["playground", "data", "strategy"]
+__all__ = ["data", "playground", "strategy"]
